\name{ISIPTA-package}
\alias{ISIPTA-package}
\docType{package}
\title{
ISIPTA network analysis
}
\description{
A Network Analysis of the Imprecise Probability Community
based on ISIPTA Electronic Proceedings (see http://www.sipta.org/)
1999-2011.
}
\details{
\tabular{ll}{
Package: \tab ISIPTA\cr
Type: \tab Package\cr
Version: \tab 0.1-6\cr
Date: \tab 2011-10-24\cr
Depends: \tab igraph, colorspace, geosphere, rworldmap\cr
Suggests: \tab XML, stringr\cr
License: \tab GPL-2\cr
}

Data set index:
\preformatted{
authors_locations       Estimated location of the authors
conferences             ISIPTA conference facts
papers                  Titles of the ISIPTA papers
papers_authors          Authors of the ISIPTA papers
papers_keywords         Keywords of the ISIPTA papers
}

See demos for an exemplary analysis of these data sets.
}
\author{
Manuel J. A. Eugster <manuel.eugster@stat.uni-muenchen.de>,
Gero Walter <gero.walter@stat.uni-muenchen.de>,
Thomas Augustin <thomas@stat.uni-muenchen.de>

Maintainer: Manuel J. A. Eugster <manuel.eugster@stat.uni-muenchen.de>
}
\keyword{package}
