% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ISIPTA_author}
\alias{ISIPTA_author}
\alias{print.ISIPTA_author}
\title{Summarize authors within the ISIPTA proceedings}
\usage{
ISIPTA_author(node)

\method{print}{ISIPTA_author}(x, show.papers = FALSE, ...)
}
\arguments{
\item{node}{character of the pattern of an author name or
integer of internal node id. It will be passed on to
\code{findSingleNode}.}

\item{x}{object of class \code{ISIPTA_author}.}

\item{show.papers}{logical, indicating whether or not paper
details should be displayed.}

\item{...}{further arguments, currently ignored.}
}
\value{
\code{ISIPTA_author} returns a list with information on
  the author of class \code{ISIPTA_author};
  \code{print.ISIPTA_author} nicely prints the result.
}
\description{
Summarizing the information about an ISIPTA author, containing
coauthors and contributed papers and conferences contributed to.
}
\examples{
sumGero <- ISIPTA_author("Gero Walter")
sumGero

## Showing detailed information of papers too
print(sumGero, show.papers = TRUE)

}
\seealso{
\code{\link{findSingleNode}},
  \code{\link{ISIPTA_coauthors}},
  \code{\link{ISIPTA_path}}
}
