% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ISIPTA_CACHE}
\alias{ISIPTA_CACHE}
\alias{ISIPTA_cache_demo}
\alias{getCached}
\alias{isCached}
\alias{setCached}
\title{Caching mechanism}
\usage{
ISIPTA_cache_demo()

getCached(name)

isCached(name)

setCached(name, value)
}
\arguments{
\item{name}{an object name to look for 
in ISIPTA cache, given as character.}

\item{value}{a value to be assigned to \code{name}.}
}
\value{
\code{ISIPTA_cache_demo} returns an environment.
  See Details.

\code{getCached} returns the object of given name
  within ISIPTA cache.

\code{isCached} returns \code{TRUE} if object
  with querried name is found within ISIPTA cache and
  else \code{FALSE}.

If there was an object previously stored
  under \code{name}, then \code{setCached} returns it
  and otherwise \code{NULL} is returned.
}
\description{
The cache is used to store all objects created with the
demos to allow for further inspection of the data. The created cache 
is used by the functions working on the collaboration network.
}
\details{
The returned environment of \code{ISIPTA_cache_demo}
  is an ancestor of the empty environment to avoid
  name clashes and contains the environment of the demos.
  Within the demo environment all object created 
  in the demos are stored.
}
\seealso{
\code{\link{ISIPTA_author}},
  \code{\link{ISIPTA_coauthors}},
  \code{\link{ISIPTA_path}}
}
