% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ISIPTA_coauthors}
\alias{ISIPTA_coauthors}
\alias{print.ISIPTA_coauthors}
\title{Summarize the coauthors}
\usage{
ISIPTA_coauthors(node, order = 1)

\method{print}{ISIPTA_coauthors}(x, ...)
}
\arguments{
\item{node}{character of the pattern of an author name or
integer of internal node id to be used as the center.
If the call to \code{\link{nodesByPattern}} or 
\code{\link{nodesById}} does not return a single node
an error is raised.}

\item{order}{non-negative numeric for the order of the
neighbourhood. It is coerced to an integer. An order of '0'
is the author herself/himself and an order of '1' returns
additionally her/his direct coauthors.}

\item{x}{object of class \code{ISIPTA_coauthors}.}

\item{...}{further arguments, currently ignored.}
}
\value{
An object of class "ISIPTA_coauthors" is a 
  \code{data.frame} containing the following components:
    \item{name}{author name}
    \item{nodeid}{respective node id}
    \item{order}{neighbourhood order}
  The rows are ordered in ascending neighbourhood order.
}
\description{
List direct coauthors for an ISIPTA author and her/his 
neighbourhood authors.
}
\examples{
## Find all coauthors of Paul Fink
ISIPTA_coauthors("Paul Fink")
  
## Find all coauthors of coauthors of Paul Fink
ISIPTA_coauthors(251, order = 2)
  
}
\seealso{
\code{\link{findSingleNode}},
  \code{\link{ISIPTA_author}},
  \code{\link{ISIPTA_path}}
}
