% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ISIPTA_path}
\alias{ISIPTA_path}
\alias{print.ISIPTA_path}
\title{Find path between authors}
\usage{
ISIPTA_path(from, to)

\method{print}{ISIPTA_path}(x, ...)
}
\arguments{
\item{from, to}{character of the pattern of an author name or
integer of internal node id to be used as the center.
If the call to \code{\link{nodesByPattern}} or 
\code{\link{nodesById}} does not return a single node
an error is raised.}

\item{x}{object of class \code{ISIPTA_path}.}

\item{...}{further arguments, currently ignored.}
}
\value{
An object of class "ISIPTA_path" is a 
  \code{data.frame} containing the following components:
    \item{name}{author name}
    \item{nodeid}{respective node id}
  The rows are ordered in the way that first row is
  'from' author and last row is 'to' author.
}
\description{
Display the path between two authors in the ISIPTA collaboration
network.
}
\details{
In case there exists no path in the network between the 
  authors in 'from' and 'to' then a warning is issued and the
  returned \code{data.frame} contains zero rows.
}
\examples{
## Path from author 251 (Paul Fink) to Gero Walter (117)
ISIPTA_path(from = 251, to = "Gero Walter")
  
}
\seealso{
\code{\link{findSingleNode}},
  \code{\link{ISIPTA_author}},
  \code{\link{ISIPTA_coauthors}}
}
