% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{nodesById}
\alias{nodesById}
\alias{nodesByPattern}
\alias{findSingleNode}
\title{Find authors within the ISIPTA collaboration network}
\usage{
nodesById(id)

nodesByPattern(pattern)

findSingleNode(node)
}
\arguments{
\item{id}{integer, node ID as shown in the graph (see
\code{demo("coauthors-network")}).}

\item{pattern}{character, containing patten of an author name. If
\code{length(pattern) > 1} only the first element is used.}

\item{node}{character of the pattern of an author name or
integer of internal node id.}
}
\value{
\code{nodesById} returns a \code{data.frame} with the
  corresponding author names. If no matching node is found,
  the returned \code{data.frame} has 0 rows.

\code{nodesByPattern} returns a \code{data.frame} with each
  author matching the \code{pattern} and the corresponding graph
  node id (see \code{demo("coauthors-network")}. If no matching
  node is found, the returned \code{data.frame} has 0 rows.

\code{findSingleNode} returns a \code{data.frame} 
  with variable names \code{name} and \code{nodeid} holding
  the author name and the corresponding graph node id.
  If the call to \code{nodesByPattern} or 
  \code{nodesById} does not return a single node
  an error is raised.
}
\description{
\code{nodesById} looks for matching nodes by ID, 
\code{nodesByPattern} by name pattern and 
\code{findSingleNode} for a single one by means of the previous
functions.
}
\details{
In contrast to \code{nodesByPattern} and
  \code{nodesById} an object is only returned if the
  input \code{node} to \code{findSingleNode} uniquely matches
  a node in the graph.
}
\examples{
## Find author by internal node id
nodesById(81:83)

## Find author by (partial) name
nodesByPattern("Paul")

## Find only one author by (partial) name
findSingleNode("Fink")

## Find only one author by nodeid
findSingleNode(251)

}
\seealso{
\code{\link{ISIPTA_author}} for looking up
  information about an author
}
