% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{authors_locations}
\alias{authors_locations}
\title{(Estimated) location of authors}
\format{A data frame with 729 observations on the following 10 variables,
where each row contains information on one author in a certain year.
\describe{
 \item{\code{year}}{Integer vector of the year of the }
 \item{\code{author}}{A factor of the authors' names.}
 \item{\code{email}}{A character vector with the authors' 
       email adress, if available.}
 \item{\code{country_code}}{A factor of the country code 
       according to ISO 3166 alpha 2.}
 \item{\code{country_name}}{A factor of the country name.}
 \item{\code{city}}{A factor of the city name.}
 \item{\code{city_lat}}{A numeric vector with the city's latitude.}
 \item{\code{city_lon}}{A numeric vector with the city's longitude.}
 \item{\code{university}}{A character vector containing the author's 
       affiliated university. Currently contains only NA}
 \item{\code{department}}{A character vector containing the author's
       affiliated department within the university.
       Currently contains only NA}
}}
\source{
Information scraped from the ISIPTA websites \url{http://www.sipta.org/}.

For details of generation see 
\url{https://github.com/paul-fink/ISIPTA/tree/master/package/generation}
and its subdirectories.
}
\usage{
data(authors_locations)
}
\description{
Location information about the ISIPTA authors.
\cr
The geocodes are obtained from the Data Science Toolkit collection 
by the help of the \code{geocode} function of the \code{ggmap} package.
}
\details{
The author names are unique within a conference, i.e. within same year, 
but are not unique overall to account for changing universities.
}
\examples{
data(authors_locations)
}
\references{
Kahle, D. and Wickham, H. (2013)  ggmap: Spatial Visualization with ggplot2.
\emph{The R Journal}, \bold{5}(1), 144--161. 
\url{https://journal.r-project.org/archive/2013-1/kahle-wickham.pdf}
}
\seealso{
\code{\link{papers_authors}}
}
\keyword{datasets}
