% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{conferences}
\alias{conferences}
\title{ISIPTA conference facts}
\format{A data frame with 10 observations on the following 10 variables.
\describe{
 \item{\code{year}}{An integer vector with the year of the conference}
 \item{\code{country_code}}{A factor of the country code according 
       to ISO 3166 alpha 2.}
 \item{\code{country_name}}{A factor of the country name.}
 \item{\code{city}}{A character vector of the city name.}
 \item{\code{city_lat}}{A numeric vector with the city's latitude.}
 \item{\code{city_lon}}{A numeric vector with the city's longitude.}
 \item{\code{university}}{A character vector containing the name of 
       the hosting university.}
 \item{\code{department}}{A character vector containing the department
       within the hosting university.}
 \item{\code{date_start}}{The start date as Date.}
 \item{\code{date_end}}{The end date as Date.}
}}
\source{
Information scraped from the ISIPTA websites \url{http://www.sipta.org/}.

For details of generation see 
\url{https://github.com/paul-fink/ISIPTA/tree/master/package/generation}
and its subdirectories.
}
\usage{
data(conferences)
}
\description{
Dates, locations, etc. about the ISIPTA conferences.
}
\examples{
data(conferences)
# Years
conferences$year
}
\seealso{
\code{\link{papers}}
}
\keyword{datasets}
