% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{papers}
\alias{papers}
\title{ISIPTA papers}
\format{A data frame with 424 observations on the following 5 variables,
where each row contains information about one paper.
\describe{
 \item{\code{year}}{The year in which the paper featured as numeric vector}
 \item{\code{id}}{A numeric vector as the unique paper id.
       The first 4 digits give the year and the last 3 the unique 
       identifier within the year.}
 \item{\code{title}}{The paper's title as a character vector}
 \item{\code{abstract}}{The paper's abstract as a character vector}
 \item{\code{url}}{The url for the paper's document in Portable Document
       Format (PDF) or PostScript (PS) as a character vector}
}}
\source{
Information scraped from the ISIPTA websites \url{http://www.sipta.org/}.

For details of generation see 
\url{https://github.com/paul-fink/ISIPTA/tree/master/package/generation}
and its subdirectories.
}
\usage{
data(papers)
}
\description{
Information about the ISIPTA papers.
}
\details{
The authors and keywords of a paper are in the dedicated objects
\code{\link{papers_authors}} and \code{\link{papers_keywords}}, 
respectively.
}
\examples{
data(papers)
# The title of the author's first paper
papers[papers$id == 2013014,]$title
}
\seealso{
\code{\link{papers_authors}},
\code{\link{authors_locations}}, 
\code{\link{papers_keywords}}
}
\keyword{datasets}
