\name{papers_authors}
\alias{papers_authors}
\docType{data}
\title{Co-authoring}
\description{
Co-authoring of ISIPTA papers.
}
\usage{data(papers_authors)}
\format{
  A data frame with 517 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{a numeric vector}
    \item{\code{id}}{a numeric vector}
    \item{\code{author}}{a factor with levels \code{Alain Chateauneuf} \code{Alane Alves  Silva} \code{Alberto Piatti} \code{Alessandro Antonucci} \code{Alessio Benavoli} \code{Alexander Karlsson} \code{Alexander Lepskiy} \code{Alfonso Suarez-Llorens} \code{Anbu Yue} \code{Andrea Sgarro} \code{Andres Cano} \code{Angelo Gilio} \code{Annette Peters} \code{Antoine Billot} \code{Anton Wallner} \code{Antonio Salmerón} \code{Arkadii Slinko} \code{Aron Larsson} \code{Astride Aregui} \code{Barbara Vantaggi} \code{Bernard De Baets} \code{Bernhard Schmelzer} \code{Brett Houlding} \code{Brigitte Jaumard} \code{Carolin Strobl} \code{Caroline Lynch} \code{Cassio Campos} \code{Cedric Baudrit} \code{Charles F. Manski} \code{Charles Geyer} \code{Charles Manski} \code{Christofer Waldenstrom} \code{Cliff Joslyn} \code{Costanza Torricelli} \code{Damjan Skulj} \code{Dan Berleant} \code{Darryl Morrell} \code{David Harmanec} \code{David Mercier} \code{David Schmeidler} \code{David Sundgren} \code{David V. Budescu} \code{Didier Dubois} \code{Dieter Denneberg} \code{Dominic Searles} \code{Don Barry} \code{Donald Bamber} \code{Elena Tsiporkova} \code{Elias Moreno} \code{Elmar Kriegler} \code{Emad Saad} \code{Enrique Miranda} \code{Eran Hanany} \code{Eric Chojnacki} \code{Eric Lefevre} \code{Erik Quaeghebeur} \code{Erio Castagnoli} \code{Etienne E. Kerre} \code{Fabien Chauny} \code{Fabio  Cozman} \code{Fabio Cuzzolin} \code{Fabio Gagliardi Cozman} \code{Fabio Maccheroni} \code{Fabio Trojani} \code{Felipe Trevizan} \code{Fernando Campello de Souza} \code{Fernando Reche} \code{Filip Hermans} \code{Francesca Vattari} \code{Frank  Coolen} \code{Frank Coolen} \code{Frank Hampel} \code{Frans Voorbraak} \code{Frederic Pichon} \code{Frederique Fournier} \code{Fulvio Fontini} \code{George J. Klir} \code{George Wilmers} \code{Gerald Sheble} \code{Gerd Peschl} \code{Gernot Kleiter} \code{Gero Walter} \code{Gert de Cooman} \code{Gert De Cooman} \code{Gertrudes Nadler Lins} \code{Giorgio Corani} \code{Giuliana Regoli} \code{Giulianella Coletti} \code{Giuseppe Busanello} \code{Giuseppe Sanfilippo} \code{Glen Meeden} \code{Glenn Shafer} \code{Guangtao Fu} \code{Henri Prade} \code{Hermann Held} \code{Horacio Arlo-Costa} \code{Hugo J. Janssen} \code{Hugo Janssen} \code{Huguette Reynaerts} \code{Hutter Marcus} \code{I.R. Goodman} \code{Ian Vernon} \code{Igor Kozine} \code{Ines Couso} \code{Ines Gonzalez-Rodriguez} \code{Isaac Levi} \code{Itzack Gilboa} \code{Itzhak Gilboa} \code{Jacinto Martín} \code{Jaime Ide} \code{James M. Dickey} \code{Javier Hernández} \code{Jean-Christophe Vergnaud} \code{Jean-Jacques Royer} \code{Jean-Marc Bernard} \code{Jean-Marc Tallon} \code{Jean-Philippe Argaud} \code{Jean-Yves Jaffray} \code{Jeff Paris} \code{Jeffrey Helzner} \code{Jianzhong Zhang} \code{Jim  Hall} \code{Jim Hall} \code{Jim Johansson} \code{Jioina Vejnarova} \code{Joaquin Abellan} \code{Joerg Stoye} \code{John Kenney} \code{Jonathan Lawry} \code{José Pablo Arias} \code{Jose Rocha} \code{Joseph B. Kadane} \code{Joseph Kadane} \code{Julian King} \code{Karen M. Kramer} \code{Karina  Delgado} \code{KazuhisaTakemura} \code{Ke-Jian Yan} \code{Kevin Loquin} \code{Klaus Nehring} \code{Kurt Weichselberger} \code{Leandro Rego} \code{Leliane Barros} \code{Lev Utkin} \code{Lev V. Utkin} \code{Love Ekenberg} \code{Luciano Sanchez} \code{Luis Raul Pericchi} \code{Malcolm Farrow} \code{Manuel Gomez} \code{Marcello Basili} \code{Marcello Mastroleo} \code{Marco Baioletti} \code{Marco Cattaneo} \code{Marco Zaffalon} \code{Marcus Hutter} \code{Marcus Poggi de Aragão} \code{Maria Eglee Perez} \code{Mark Ducey} \code{Mark J. Schervish} \code{Mark Schervish} \code{Marston Conder} \code{Massimo Marinacci} \code{Mathieu Dancre} \code{Mats Danielson} \code{Matthias Troffaes} \code{Meglena   Jeleva} \code{Meglena Jeleva} \code{Michèle  Vanmaele} \code{Michael Goldstein} \code{Michael Oberguggenberger} \code{Michael Obermeier} \code{Michael Smithson} \code{Michel Grabisch} \code{Michele Cohen} \code{Mikael Andersson} \code{Mikelis Bickis} \code{Milan Daniel} \code{Minh Ha Duong} \code{Morgane Cheve} \code{Nathan Huntley} \code{Nic Wilson} \code{Niki Pfeifer} \code{Olivier Strauss} \code{P.N.Watton} \code{Pablo Fierens} \code{Pablo Ignacio Fierens} \code{Paola Castellan} \code{Paolo Ghirardato} \code{Paolo Vicig} \code{Paul Snow} \code{Pauline Coolen-Schrijner} \code{Pedro Gil} \code{Pei Wang} \code{Peter Gillett} \code{Peter Harremoes} \code{Peter Klibanoff} \code{Peter P. Wakker} \code{Peter Walley} \code{Philippe Nivlet} \code{Philippe Smets} \code{Pierre Hansen} \code{Prakash P. Shenoy} \code{Radim Jirousek} \code{Radu Lazar} \code{Ralph Bruehlmann} \code{Raphael Giraud} \code{Rebecca Baker} \code{Renato Pelessoni} \code{Ricardo Shirota Filho} \code{Richard Crossman} \code{Richard Scherl} \code{Robert F. Nau} \code{Robert Hable} \code{Robert Nau} \code{Robert Winkler} \code{Robin Christian} \code{Rolf Haenni} \code{Romano Scozzafava} \code{Ronald Yager} \code{Ronan Congar} \code{Ronnie Johansson} \code{Ruggero Frezza} \code{Salem Benferhat} \code{Salma Smaoui} \code{Sebastian Maass} \code{Serafin Moral} \code{Serena Doria} \code{Sergey V. Gurov} \code{Silvia Muzzioli} \code{Simon Marshall} \code{Simon Pittschmann} \code{Stefan Arnborg} \code{Sten F Andler} \code{Steven Parkinson} \code{Sujoy Mukerji} \code{Sun Yi} \code{Susana Montes} \code{Sven-Hendrik Lossin} \code{Svetlana Zatenko} \code{Sylvain Perron} \code{Tahani Maturi} \code{Teddy Seidenfeld} \code{Terrence Fine} \code{Thierry Denoeux} \code{Thomas Augustin} \code{Thomas Bartos} \code{Thomas Fetz} \code{Thomas Lukasiewicz} \code{Tomas Kroupa} \code{Tzur Karelitz} \code{Ugis Bickis} \code{Veronica Biazzo} \code{Veselka Boeva} \code{Victor Krymsky} \code{Victor Richmond  Jose} \code{Vincent Fortin} \code{Vladislav Bina} \code{Volker Kraetschmer} \code{Weiru Liu} \code{Wynn Stirling} \code{Yakov Ben-Haim} \code{Yongchuan Tang}}
  }
}
\source{
Screen scraped from the ISIPTA website \url{http://www.sipta.org/}; see directories \code{scraper} and \code{xml} for the screen scrapers and the raw data.
}
\examples{
data(papers_authors)
## maybe str(papers_authors) ; plot(papers_authors) ...
}
\keyword{datasets}
\seealso{
\code{\link{papers}}, \code{\link{authors_locations}}, \code{\link{papers_keywords}}
}
