\name{papers_authors}
\alias{papers_authors}
\docType{data}
\title{Authors of the ISIPTA papers}
\description{
Authors of the ISIPTA papers; the names are normalized to ASCII characters.
}
\usage{data("papers_authors")}
\format{
  A data frame with 689 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{a numeric vector}
    \item{\code{id}}{a numeric vector}
    \item{\code{author}}{a factor with levels \code{Alain Chateauneuf} \code{Alain Vandal} \code{Alane Alves Silva} \code{Alberto Piatti} \code{Alessandro Antonucci} \code{Alessandro Giusti} \code{Alessio Benavoli} \code{Alexander Karlsson} \code{Alexander Lepskiy} \code{Alfonso Suarez-Llorens} \code{Alice Cicirello} \code{Anbu Yue} \code{Andrea Capotorti} \code{Andrea Mignatti} \code{Andrea Sgarro} \code{Andrea Wiencierz} \code{Andres Cano} \code{Andres Masegosa} \code{Andrew Bronevich} \code{Angelo Gilio} \code{Anne Kunz} \code{Annette Peters} \code{Antoine Billot} \code{Antonio Salmeron} \code{Anton Wallner} \code{Arkadii Slinko} \code{Aron Larsson} \code{Arthur van Camp} \code{Astride Aregui} \code{Barbara Vantaggi} \code{Bernard De Baets} \code{Bernhard Schmelzer} \code{Brett Houlding} \code{Brigitte Jaumard} \code{Caroline Lynch} \code{Carolin Strobl} \code{Cassio Campos} \code{Cedric Baudrit} \code{Charles Geyer} \code{Charles Manski} \code{Christelle Jacob} \code{Christofer Waldenstrom} \code{Christophe Berenguer} \code{Cleotilde Gonzalez} \code{Cliff Joslyn} \code{Costanza Torricelli} \code{Damjan Skulj} \code{Dan Berleant} \code{Daniel Jolly} \code{Darryl Morrell} \code{David Budescu} \code{Davide Petturiti} \code{David Harmanec} \code{David Mercier} \code{David Schmeidler} \code{David Sundgren} \code{Denis Maua} \code{Didier Dubois} \code{Dieter Denneberg} \code{Diogo Bezerra} \code{Dominic Searles} \code{Dominique Vasseur} \code{Donald Bamber} \code{Don Barry} \code{Elena Tsiporkova} \code{Elias Moreno} \code{Elmar Kriegler} \code{Emad Saad} \code{Enrique Miranda} \code{Eran Hanany} \code{Eric Chojnacki} \code{Eric Lefevre} \code{Erik Quaeghebeur} \code{Erio Castagnoli} \code{Etienne E. Kerre} \code{Fabien Chauny} \code{Fabio Cozman} \code{Fabio Cuzzolin} \code{Fabio Maccheroni} \code{Fabio Trojani} \code{Felipe Aguirre} \code{Felipe Trevizan} \code{Fernando Campello de Souza} \code{Fernando Reche} \code{Filip Hermans} \code{Francesca Mangili} \code{Francesca Vattari} \code{Frank Coolen} \code{Frank Hampel} \code{Frans Voorbraak} \code{Frederic Pichon} \code{Frederique Fournier} \code{Fulvio Fontini} \code{George Klir} \code{George Wilmers} \code{Georg Schollmeyer} \code{Gerald Sheble} \code{Gerd Peschl} \code{Gernot Kleiter} \code{Gero Walter} \code{Gert de Cooman} \code{Gertrudes Nadler Lins} \code{Giorgio Corani} \code{Giuliana Regoli} \code{Giulianella Coletti} \code{Giuseppe Busanello} \code{Giuseppe Sanfilippo} \code{Glen Meeden} \code{Glenn Shafer} \code{Guangtao Fu} \code{Hailin Liu} \code{Han-Hui Por} \code{Helmut Kuechenhoff} \code{Henri Prade} \code{Hermann Held} \code{Horacio Arlo-Costa} \code{Hugo Janssen} \code{Huguette Reynaerts} \code{Hutter Marcus} \code{Ian Vernon} \code{Ignacio Montes} \code{Igor Kozine} \code{Igor Rozenberg} \code{Ines Couso} \code{Ines Gonzalez-Rodriguez} \code{Irwin Goodman} \code{Isaac Levi} \code{Itzack Gilboa} \code{Itzhak Gilboa} \code{Jacinto Martin} \code{Jack Siegrist} \code{Jaime Ide} \code{James Dickey} \code{Jasper de Bock} \code{Javier Hernandez} \code{Jean-Christophe Vergnaud} \code{Jean-Jacques Royer} \code{Jean-Marc Bernard} \code{Jean-Marc Tallon} \code{Jean-Philippe Argaud} \code{Jean-Yves Jaffray} \code{Jeff Paris} \code{Jeffrey Helzner} \code{Jianzhong Zhang} \code{Jim Hall} \code{Jim Johansson} \code{Jirina Vejnarova} \code{Joaquin Abellan} \code{Joerg Stoye} \code{Joe Steinhauer} \code{John Kenney} \code{John Paul Gosling} \code{Jonathan Lawry} \code{Jose Pablo Arias} \code{Joseph Kadane} \code{Jose Rocha} \code{Julian King} \code{Kamal Premaratne} \code{Karen Kramer} \code{Karina Delgado} \code{Kari Sentz} \code{Kazuhisa Takemura} \code{Ke-Jian Yan} \code{Kevin Loquin} \code{Klaus Nehring} \code{Kurt Weichselberger} \code{Laurence Dieulle} \code{Leandro Rego} \code{Leliane Barros} \code{Lev Utkin} \code{Lewis Paton} \code{Love Ekenberg} \code{Luciano Sanchez} \code{Luis Raul Pericchi} \code{Malcolm Farrow} \code{Manohar N. Murthi} \code{Manuel Gomez} \code{Marcello Basili} \code{Marcello Mastroleo} \code{Marco Baioletti} \code{Marco Cattaneo} \code{Marco Zaffalon} \code{Marcus Hutter} \code{Marcus Poggi de Aragao} \code{Maria Eglee Perez} \code{Mark Ducey} \code{Mark Schervish} \code{Marston Conder} \code{Massimo Marinacci} \code{Mathieu Couplet} \code{Mathieu Dancre} \code{Mats Danielson} \code{Matthias Scheutz} \code{Matthias Troffaes} \code{Matthieu Hourbracq} \code{Meglena Jeleva} \code{Michael Balch} \code{Michael Beer} \code{Michael Goldstein} \code{Michael Oberguggenberger} \code{Michael Obermeier} \code{Michael Smithson} \code{Michele Cohen} \code{Michele Vanmaele} \code{Michel Grabisch} \code{Mikael Andersson} \code{Mikelis Bickis} \code{Milan Daniel} \code{Mingqiang Zhang} \code{Minh Ha Duong} \code{Mohamed Sallak} \code{Morgane Cheve} \code{Nathan Huntley} \code{Nic Wilson} \code{Niki Pfeifer} \code{Olivier Strauss} \code{Pablo Fierens} \code{Paola Castellan} \code{Paolo Ghirardato} \code{Paolo Vicig} \code{Paul Fink} \code{Pauline Coolen-Schrijner} \code{Paul Snow} \code{Paul Watton} \code{Pedro Gil} \code{Pei Wang} \code{Peter Gillett} \code{Peter Harremoes} \code{Peter Klibanoff} \code{Peter Wakker} \code{Peter Walley} \code{Philippe Nivlet} \code{Philippe Smets} \code{Pierre Hansen} \code{Pierre-Henri Wuillemin} \code{Prakash Shenoy} \code{Radim Jirousek} \code{Radu Lazar} \code{Rafael Nunez} \code{Ralph Bruehlmann} \code{Raphael Giraud} \code{Rebecca Baker} \code{Renato Pelessoni} \code{Ricardo Shirota Filho} \code{Richard Crossman} \code{Richard Scherl} \code{Robert Hable} \code{Robert Nau} \code{Robert Winkler} \code{Robin Christian} \code{Robin Langley} \code{Rocco De Rosa} \code{Rolf Haenni} \code{Romano Scozzafava} \code{Ronald Yager} \code{Ronan Congar} \code{Ronnie Johansson} \code{Ruggero Frezza} \code{Salem Benferhat} \code{Salma Smaoui} \code{Scott Ferson} \code{Sebastian Maass} \code{Sebastien Destercke} \code{Serafin Moral} \code{Serena Doria} \code{Sergey Gurov} \code{Ser Tong Quek} \code{Shoumei Li} \code{Silvia Muzzioli} \code{Simon Blake} \code{Simon Marshall} \code{Simon Pittschmann} \code{Stefan Arnborg} \code{Sten Andler} \code{Stephen Broomell} \code{Steven Parkinson} \code{Sujoy Mukerji} \code{Sun Yi} \code{Susana Montes} \code{Sven-Hendrik Lossin} \code{Svetlana Zatenko} \code{Sylvain Perron} \code{Tahani Maturi} \code{Teddy Seidenfeld} \code{Terrence Fine} \code{Thierry Denoeux} \code{Thomas Augustin} \code{Thomas Bartos} \code{Thomas Fetz} \code{Thomas Lukasiewicz} \code{Tomas Kroupa} \code{Tu Duong LE DUY} \code{Tzur Karelitz} \code{Ugis Bickis} \code{Vaclav Kratochvil} \code{Varun Dutt} \code{Veronica Biazzo} \code{Veselka Boeva} \code{Victor Krymsky} \code{Victor Richmond Jose} \code{Vincent Fortin} \code{Vladislav Bina} \code{Volker Kraetschmer} \code{Weiru Liu} \code{Wynn Stirling} \code{Xuecheng Liu} \code{Xun Wang} \code{Yakov Ben-Haim} \code{Yongchuan Tang}}
  }
}
\source{
Screen scraped from the ISIPTA website \url{http://www.sipta.org/}; see directories \code{scraper} and \code{xml} for the screen scrapers and the raw data.
}
\examples{
data(papers_authors)
## maybe str(papers_authors) ; plot(papers_authors) ...
}
\keyword{datasets}
\seealso{
\code{\link{papers}}, \code{\link{authors_locations}}, \code{\link{papers_keywords}}
}
