% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{papers_authors}
\alias{papers_authors}
\title{Authors of the ISIPTA papers}
\format{A data frame with 848 observations on the following 3 variables,
where each row contains one author of a paper.
\describe{
 \item{\code{year}}{The year in which the paper featured as numeric vector.}
 \item{\code{id}}{A numeric vector as the unique paper id.
       See \code{\link{papers}} for details.}
 \item{\code{author}}{A factor with name of the papers' authors.
       The names are normalized to ASCII characters.}
}}
\source{
Information scraped from the ISIPTA websites \url{http://www.sipta.org/}.

For details of generation see 
\url{https://github.com/paul-fink/ISIPTA/tree/master/package/generation}
and its subdirectories.
}
\usage{
data(papers_authors)
}
\description{
Matching authors to the ISIPTA papers
}
\examples{
data(papers_authors)

# Co-authors for paper with id 2013014
papers_authors[papers_authors$id == 2013014,]

# Table of all authors
table(papers_authors$author)
}
\seealso{
\code{\link{papers}},
\code{\link{authors_locations}}, 
\code{\link{papers_keywords}}
}
\keyword{datasets}
