% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{papers_keywords}
\alias{papers_keywords}
\title{Keywords of the ISIPTA papers}
\format{A data frame with 2184 observations on the following 3 variables,
where each row contains one keyword of a paper.
\describe{
 \item{\code{year}}{The year in which the paper featured as numeric vector.}
 \item{\code{id}}{A numeric vector as the unique paper id.
       See \code{\link{papers}} for details.}
 \item{\code{keyword}}{A character vector with the papers' keywords.
       The keywords are slightly normalized.}
}}
\source{
Information scraped from the ISIPTA websites \url{http://www.sipta.org/}.

For details of generation see 
\url{https://github.com/paul-fink/ISIPTA/tree/master/package/generation}
and its subdirectories.
}
\usage{
data(papers_keywords)
}
\description{
Keywords of the ISIPTA papers.
}
\examples{
data(papers_keywords)
}
\seealso{
\code{\link{papers}},
\code{\link{authors_locations}}, 
\code{\link{papers_authors}}
}
\keyword{datasets}
